Profil apoteke
Uradjen pregled profila apoteke, ima bag pri azuriranju

+++ ISAMRS_PROJEKAT/src/main/java/rest/domain/Apoteka.java
@@ -9,7 +9,7 @@ public class Apoteka implements Ocenjivo{
 	private String opis;
 	private int brojOcena;
 	private int sumaOcena;
-	private int ocena;
+	private double ocena;
 	
 	private Set<Pacijent> pacijenti;
 	private Set<Pregled> pregledi;
@@ -37,7 +37,7 @@ public class Apoteka implements Ocenjivo{
 		this.ocena = this.izracunajOcenu();
 	}
 	
-	public int izracunajOcenu()
+	public double izracunajOcenu()
 	{
 		if(this.brojOcena!=0) {
 			return this.sumaOcena/this.brojOcena;
@@ -133,11 +133,11 @@ public class Apoteka implements Ocenjivo{
 		this.adminiApoteke = adminiApoteke;
 	}
 
-	public int getOcena() {
+	public double getOcena() {
 		return ocena;
 	}
 
-	public void setOcena(int ocena) {
+	public void setOcena(double ocena) {
 		this.ocena = ocena;
 	}
