+++ ISAMRS_PROJEKAT/src/main/java/rest/domain/Farmaceut.java
@@ -19,7 +19,7 @@ public class Farmaceut extends Korisnik implements Ocenjivo{
 		this.ocena=izracunajOcenu();
 	}
 
-	public int izracunajOcenu()
+	public double izracunajOcenu()
 	{
 		if(this.brojOcena!=0) {
 			return this.sumaOcena/this.brojOcena;
