+++ ISAMRS_PROJEKAT/src/main/java/rest/domain/Ocenjivo.java
@@ -2,5 +2,5 @@ package rest.domain;
 
 public interface Ocenjivo {
 
-	public int izracunajOcenu();
+	public double izracunajOcenu();
 }
